Comment operand.rs
