+++ kani-compiler/src/codegen_cprover_gotoc/codegen/place.rs
@@ -310,7 +310,7 @@ impl<'tcx> GotocCtx<'tcx> {
     /// a named variable.
     ///
     /// Recursively finds the actual FnDef from a pointer or box.
-    pub fn codegen_local_fndef(&mut self, ty: ty::Ty<'tcx>) -> Option<Expr> {
+    fn codegen_local_fndef(&mut self, ty: ty::Ty<'tcx>) -> Option<Expr> {
         match ty.kind() {
             // A local that is itself a FnDef, like Fn::call_once
             ty::FnDef(defid, substs) => Some(self.codegen_fndef(*defid, substs, None)),
@@ -329,7 +329,7 @@ impl<'tcx> GotocCtx<'tcx> {
     }
 
     /// Codegen for a local
-    pub fn codegen_local(&mut self, l: Local) -> Expr {
+    fn codegen_local(&mut self, l: Local) -> Expr {
         // Check if the local is a function definition (see comment above)
         if let Some(fn_def) = self.codegen_local_fndef(self.local_ty(l)) {
             return fn_def;
